Add support for downloads
A link which triggers a download will take the file and put it where the
PDF of the page is expected. If the download is a PDF then this will end
up with the PDF merged into the result. If the download isn't a PDF
then... things won't really work!

TODO:
 - Work out why the runtime prints a warning about an exception from the
   download future which isn't retrieved (the page close exception)
 - Check that the downloaded file looks like a PDF